+++ src/pydvl/utils/dataset.py
@@ -68,9 +68,8 @@ class Dataset:
             timestamp which can be referenced directly instead of using a row
             number.
         :param description: A textual description of the dataset.
-        :param is_multi_output: set to True if y holds multiple labels for each
-            data point. False if y is a 1d array holding a single label per
-            point.
+        :param is_multi_output: set to ``False`` if labels are scalars, or to
+            ``True`` if they are vectors of dimension > 1.
         """
         self.x_train, self.y_train = check_X_y(
             x_train, y_train, multi_output=is_multi_output
@@ -130,9 +129,6 @@ class Dataset:
         self.description = description or "No description"
         self._indices = np.arange(len(self.x_train))
         self._data_names = data_names if data_names is not None else self._indices
-
-    def __iter__(self):
-        return self.x_train, self.y_train, self.x_test, self.y_test
 
     def __getitem__(self, idx: Union[int, slice, Iterable]) -> Tuple:
         return self.x_train[idx], self.y_train[idx]
@@ -253,22 +249,33 @@ class Dataset:
         train_size: float = 0.8,
         random_state: Optional[int] = None,
         stratify_by_target: bool = False,
+        **kwargs,
     ) -> "Dataset":
-        """Constructs a :class:`Dataset` object from an :class:`sklearn.utils.Bunch` bunch as returned by the
-        `load_*` functions in `sklearn toy datasets
+        """Constructs a :class:`Dataset` object from an
+        :class:`sklearn.utils.Bunch`, as returned by the `load_*` functions in
+        `sklearn toy datasets
         <https://scikit-learn.org/stable/datasets/toy_dataset.html>`_.
 
-        :param data: sklearn dataset
+        :param data: sklearn dataset. The following attributes are supported
+            - ``data``: covariates [required]
+            - ``target``: target variables (labels) [required]
+            - ``feature_names``: the feature names
+            - ``target_names``: the target names
+            - ``DESCR``: a description
         :param train_size: size of the training dataset. Used in
             `train_test_split`
         :param random_state: seed for train / test split
         :param stratify_by_target: If `True`, data is split in a stratified
             fashion, using the target variable as labels. Read more in
-            `sklearn's user guide
+            `scikit-learn's user guide
             <https://scikit-learn.org/stable/modules/cross_validation.html
             #stratification>`.
+        :param kwargs: Additional keyword arguments to pass to the
+            :class:`Dataset` constructor. Use this to pass e.g. ``is_multi_output``.
+        :return: Object with the sklearn dataset
 
-        :return: Dataset with the selected sklearn data
+        .. versionchanged:: 0.6.0
+           Added kwargs to pass to the :class:`Dataset` constructor.
         """
         x_train, x_test, y_train, y_test = train_test_split(
             data.data,
@@ -285,6 +292,7 @@ class Dataset:
             feature_names=data.get("feature_names"),
             target_names=data.get("target_names"),
             description=data.get("DESCR"),
+            **kwargs,
         )
 
     @classmethod
@@ -295,11 +303,10 @@ class Dataset:
         train_size: float = 0.8,
         random_state: Optional[int] = None,
         stratify_by_target: bool = False,
+        **kwargs,
     ) -> "Dataset":
-        """.. versionadded:: 0.4.0
-
-        Constructs a :class:`Dataset` object from X and y numpy arrays  as returned by the
-        `make_*` functions in `sklearn generated datasets
+        """Constructs a :class:`Dataset` object from X and y numpy arrays  as
+        returned by the `make_*` functions in `sklearn generated datasets
         <https://scikit-learn.org/stable/datasets/sample_generators.html>`_.
 
         :param X: numpy array of shape (n_samples, n_features)
@@ -312,8 +319,16 @@ class Dataset:
             `sklearn's user guide
             <https://scikit-learn.org/stable/modules/cross_validation.html
             #stratification>`.
+        :param kwargs: Additional keyword arguments to pass to the
+            :class:`Dataset` constructor. Use this to pass e.g. ``feature_names``
+            or ``target_names``.
+        :return: Object with the passed X and y arrays split across training
+            and test sets.
 
-        :return: Dataset with the passed X and y arrays split across training and test sets.
+        .. versionadded:: 0.4.0
+
+        .. versionchanged:: 0.6.0
+           Added kwargs to pass to the :class:`Dataset` constructor.
         """
         x_train, x_test, y_train, y_test = train_test_split(
             X,
@@ -322,7 +337,7 @@ class Dataset:
             random_state=random_state,
             stratify=y if stratify_by_target else None,
         )
-        return cls(x_train, y_train, x_test, y_test)
+        return cls(x_train, y_train, x_test, y_test, **kwargs)
 
 
 class GroupedDataset(Dataset):
@@ -337,6 +352,7 @@ class GroupedDataset(Dataset):
         target_names: Optional[Sequence[str]] = None,
         group_names: Optional[Sequence[str]] = None,
         description: Optional[str] = None,
+        **kwargs,
     ):
         """Class for grouping datasets.
 
@@ -357,14 +373,22 @@ class GroupedDataset(Dataset):
         :param target_names: names of the labels or targets y
         :param group_names: names of the groups. If not provided, the labels
             from ``data_groups`` will be used.
+        :param description: A textual description of the dataset
+        :param kwargs: Additional keyword arguments to pass to the
+            :class:`Dataset` constructor.
 
-        :param description: description of the dataset
-
-        .. versionchanged:: 0.5.1
-           Added `group_names` argument
+        .. versionchanged:: 0.6.0
+           Added ``group_names`` and forwarding of ``kwargs``
         """
         super().__init__(
-            x_train, y_train, x_test, y_test, feature_names, target_names, description
+            x_train=x_train,
+            y_train=y_train,
+            x_test=x_test,
+            y_test=y_test,
+            feature_names=feature_names,
+            target_names=target_names,
+            description=description,
+            **kwargs,
         )
 
         if len(data_groups) != len(x_train):
@@ -423,32 +447,52 @@ class GroupedDataset(Dataset):
         random_state: Optional[int] = None,
         stratify_by_target: bool = False,
         data_groups: Optional[Sequence] = None,
+        **kwargs,
     ) -> "GroupedDataset":
-        """Constructs a :class:`GroupedDataset` object from an sklearn bunch as returned by the
-        `load_*` functions in `sklearn toy datasets
+        """Constructs a :class:`GroupedDataset` object from a scikit-learn bunch
+        as returned by the `load_*` functions in `sklearn toy datasets
         <https://scikit-learn.org/stable/datasets/toy_dataset.html>`_ and groups
         it.
 
-        :param data: sklearn dataset
+        :param data: sklearn dataset. The following attributes are supported
+            - ``data``: covariates [required]
+            - ``target``: target variables (labels) [required]
+            - ``feature_names``: the feature names
+            - ``target_names``: the target names
+            - ``DESCR``: a description
         :param train_size: size of the training dataset. Used in
             `train_test_split`.
         :param random_state: seed for train / test split.
-        :param stratify_by_target: If `True`, data is split in a stratified
+        :param stratify_by_target: If ``True``, data is split in a stratified
             fashion, using the target variable as labels. Read more in
             `sklearn's user guide
             <https://scikit-learn.org/stable/modules/cross_validation.html
             #stratification>`.
-        :param data_groups: for each element in the training set, it associates
-            a group index or name.
-
+        :param data_groups: an array holding the group index or name for each
+            data point. The length of this array must be equal to the number of
+            data points in the dataset.
+        :param kwargs: Additional keyword arguments to pass to the
+            :class:`Dataset` constructor.
         :return: Dataset with the selected sklearn data
         """
         if data_groups is None:
-            raise ValueError("data_groups argument is missing")
-        dataset = Dataset.from_sklearn(
-            data, train_size, random_state, stratify_by_target
+            raise ValueError(
+                "data_groups must be provided when constructing a GroupedDataset"
+            )
+
+        x_train, x_test, y_train, y_test, data_groups_train, _ = train_test_split(
+            data.data,
+            data.target,
+            data_groups,
+            train_size=train_size,
+            random_state=random_state,
+            stratify=data.target if stratify_by_target else None,
         )
-        return cls.from_dataset(dataset, data_groups)
+
+        dataset = Dataset(
+            x_train=x_train, y_train=y_train, x_test=x_test, y_test=y_test, **kwargs
+        )
+        return cls.from_dataset(dataset, data_groups_train)  # type: ignore
 
     @classmethod
     def from_arrays(
@@ -459,6 +503,7 @@ class GroupedDataset(Dataset):
         random_state: Optional[int] = None,
         stratify_by_target: bool = False,
         data_groups: Optional[Sequence] = None,
+        **kwargs,
     ) -> "Dataset":
         """Constructs a :class:`GroupedDataset` object from X and y numpy arrays
         as returned by the `make_*` functions in `sklearn generated datasets
@@ -474,32 +519,50 @@ class GroupedDataset(Dataset):
             `sklearn's user guide
             <https://scikit-learn.org/stable/modules/cross_validation.html
             #stratification>`.
-        :param data_groups: for each element in the training set, and associated
-            group index or name.
+        :param data_groups: an array holding the group index or name for each
+            data point. The length of this array must be equal to the number of
+            data points in the dataset.
+        :param kwargs: Additional keyword arguments that will be passed
+            to the :class:`~pydvl.utils.dataset.Dataset` constructor.
 
         :return: Dataset with the passed X and y arrays split across training
             and test sets.
 
         .. versionadded:: 0.4.0
+
+        .. versionchanged:: 0.6.0
+           Added kwargs to pass to the :class:`Dataset` constructor.
         """
         if data_groups is None:
-            raise ValueError("data_groups argument is missing")
-        dataset = Dataset.from_arrays(
-            X, y, train_size, random_state, stratify_by_target
+            raise ValueError(
+                "data_groups must be provided when constructing a GroupedDataset"
+            )
+        x_train, x_test, y_train, y_test, data_groups_train, _ = train_test_split(
+            X,
+            y,
+            data_groups,
+            train_size=train_size,
+            random_state=random_state,
+            stratify=y if stratify_by_target else None,
         )
-        return cls.from_dataset(dataset, data_groups)
+        dataset = Dataset(
+            x_train=x_train, y_train=y_train, x_test=x_test, y_test=y_test, **kwargs
+        )
+        return cls.from_dataset(dataset, data_groups_train)
 
     @classmethod
     def from_dataset(
         cls, dataset: Dataset, data_groups: Sequence[Any]
     ) -> "GroupedDataset":
-        """Given a :class:`Dataset` object, it creates it into a :class:`GroupedDataset` object
-        by passing a list of data groups, one for each element in the training set.
+        """Creates a :class:`GroupedDataset` object from the data a
+        :class:`Dataset` object and a mapping of data groups.
 
-        :param dataset: :class:`Dataset` object
-        :param data_groups: for each element in the training set, it associates a group
-            index or name.
-        :return: :class:`GroupedDataset`, with the initial :class:`Dataset` grouped by data_groups.
+        :param dataset: The original data.
+        :param data_groups: An array holding the group index or name for each
+            data point. The length of this array must be equal to the number of
+            data points in the dataset.
+        :return: A :class:`GroupedDataset` with the initial :class:`Dataset`
+            grouped by data_groups.
         """
         return cls(
             x_train=dataset.x_train,